fix : order ex 핸들러 클래스 중복 삭제
