+++ src/main/java/org/example/buyingserver/order/handler/OrderExceptionHandler.java
@@ -1,6 +1,7 @@
 package org.example.buyingserver.order.handler;
 
 import org.example.buyingserver.order.exception.OrderNotFoundException;
+import org.example.buyingserver.order.exception.OrderNotReadyException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.example.buyingserver.common.dto.ApiResponse;
@@ -11,6 +12,13 @@ public class OrderExceptionHandler {
 
     @ExceptionHandler(OrderNotFoundException.class)
     public ResponseEntity<ApiResponse<?>> handleOrderException(OrderNotFoundException e) {
+        return ResponseEntity
+                .status(e.getErrorCode().getStatus())
+                .body(ApiResponse.error(e.getErrorCode()));
+    }
+
+    @ExceptionHandler(OrderNotReadyException.class)
+    public ResponseEntity<ApiResponse<?>> handleOrderNotReady(OrderNotReadyException e) {
         return ResponseEntity
                 .status(e.getErrorCode().getStatus())
                 .body(ApiResponse.error(e.getErrorCode()));
